pyproject.toml
[build-system]
requires = ["poetry-core>=2.0,<3.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "agentic_rec"
version = "0.1.0"
description = "Agentic Recommenders with MovieLens"
authors = [
  { name = "YuXuan Tay", email = "5795122+yxtay@users.noreply.github.com" },
]
license = "MIT"
requires-python = ">=3.12,<4"
dependencies = [
  "fastapi[standard]>=0.136,<1.0",
  "jsonargparse[signatures]~=4.48",
  "lancedb[pylance]>=0.19,<1.0",
  "loguru>=0.7,<1.0",
  "polars[pandas]~=1.40",
  "pydantic-ai~=1.43",
  "pydantic-settings~=2.0",
  "rerankers[pylate,transformers]>=0.9,<1.0",
  "sentence-transformers~=5.3",
  "sqlalchemy~=2.0",
  "torch~=2.9",
]

[dependency-groups]
dev = [
  "huggingface-hub[cli]~=1.14",
  "pyarrow-stubs>=20.0",
  "pytest~=9.0",
  "pytest-asyncio~=1.0",
]

[project.scripts]
data = "agentic_rec.data:cli"
app = "agentic_rec.main:cli"

[tool.fastapi]
entrypoint = "agentic_rec.main:app"

[tool.uv.sources]
torch = [{ index = "torch-cpu" }]

[[tool.uv.index]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
fix = true
src = [".", "src", "tests"]
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "D",       # pydocstyle
  "E501",    # line-too-long
  "S101",    # assert
  "S311",    # suspicious-non-cryptographic-random-usage
  "T201",    # print
  "COM812",  # missing-trailing-comma, conflicts with ruff formatter
  "ERA001",  # commented-out-code
  "ISC001",  # single-line-implicit-string-concatenation, conflicts with ruff formatter
  "PLC0415", # import-outside-top-level
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR2004"]

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.ruff.lint.pydocstyle]
convention = "pep257"
